move createtoken and verifyuniqueuser to separate modules
