+++ nodeAppPostPg/token.ts
@@ -1,14 +1,19 @@
 const jwt = require('jsonwebtoken')
 const secret = process.env.secret || 'secret'
 
-const createToken=user =>{
-    let scopes;
+const createToken = user => {
+  let scopes;
 
-    if (user.admin){
-        scopes = 'admin'
-    }
+  if (user.isAdmin) {
+      scopes = 'admin';
+  }
 
-    return jwt.sign({id: user.id, username: user.username, scope: scopes}, secret, {algorithm: 'HS256', expiresIn: "1h"});
-}
-
-module.exports = createToken;
+  return jwt.sign(
+      { id: user.id, userName: user.userName, scope: scopes}, 
+      secret, 
+      {algorithm: 'HS256', expiresIn: '1h'
+      }
+  );
+};
+  
+export {createToken};
